♻️ Moved StatisticalInterval import from utils to stats module.
