+++ tinyshift/outlier/spad.py
@@ -5,7 +5,7 @@ from collections import Counter
 from sklearn.decomposition import PCA
 from .base import BaseHistogramModel
 from typing import Union
-from ..utils import StatisticalInterval
+from ..stats import StatisticalInterval
 
 
 class SPAD(BaseHistogramModel):
