+++ tinyshift/stats/__init__.py
@@ -1,2 +1,3 @@
 from .utils import chebyshev_guaranteed_percentage
 from .statistical_interval import StatisticalInterval
+from .bootstrap_bca import BootstrapBCA
