+++ tinyshift/stats/statistical_interval.py
