+++ tinyshift/tracker/anomaly.py
@@ -14,7 +14,7 @@ class AnomalyTracker(BaseModel):
         n_resamples: int = 1000,
         random_state: int = 42,
         drift_limit: Union[str, Tuple[float, float]] = "stddev",
-        enable_confidence_interval: bool = False,
+        confidence_interval: bool = False,
     ):
         """
         A tracker for monitoring anomaly over time using a specified evaluation metric.
@@ -64,7 +64,7 @@ class AnomalyTracker(BaseModel):
             n_resamples,
             random_state,
             drift_limit,
-            enable_confidence_interval,
+            confidence_interval,
         )
 
     def score(
